+++ src/skibidi_orm/migration_engine/db_inspectors/db_inspector.py
@@ -1,7 +1,7 @@
 from typing import Any, Literal, cast
 import sqlite3
 from skibidi_orm.migration_engine.adapters.sqlite3_adapter import SQLite3Adapter
-from skibidi_orm.migration_engine.config import SQLite3Config
+from skibidi_orm.migration_engine.db_config.sqlite3_config import SQLite3Config
 from skibidi_orm.migration_engine.db_inspectors.base_inspector import BaseDbInspector
